+++ src/test/java/kitchenpos/menu/ui/MenuRestControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui;
+package kitchenpos.menu.ui;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kitchenpos.common.Name;
