+++ src/test/java/kitchenpos/table/ui/TableRestControllerTest.java
@@ -1,10 +1,9 @@
-package kitchenpos.ui;
+package kitchenpos.table.ui;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kitchenpos.order.application.OrderService;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.dto.*;
-import kitchenpos.order.ui.OrderRestController;
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.table.ui.TableRestController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -15,10 +14,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -29,72 +25,77 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @ExtendWith(MockitoExtension.class)
-class OrderRestControllerTest {
-    private static final String URI = "/api/orders";
+class TableRestControllerTest {
+    private static final String URI = "/api/tables";
 
     @InjectMocks
     private ObjectMapper objectMapper;
-
     @InjectMocks
-    private OrderRestController orderRestController;
+    private TableRestController tableRestController;
 
     @Mock
-    private OrderService orderService;
+    private TableService tableService;
 
     private MockMvc mockMvc;
-    private OrderRequest 주문_요청;
-    private OrderResponse 주문_응답;
-    private List<OrderLineItemRequest> 주문_항목;
+    private OrderTableResponse 주문_테이블_응답_1;
 
     @BeforeEach
     void setUp() {
-        this.mockMvc = MockMvcBuilders.standaloneSetup(orderRestController).build();
-        주문_항목 = Arrays.asList(new OrderLineItemRequest(1L, 1));
-        주문_요청 = new OrderRequest(1L, 주문_항목);
-        주문_응답 = new OrderResponse(1L, 1L, OrderStatus.COOKING.name(), null,
-                new ArrayList<>(Arrays.asList(new OrderLineItemResponse(1L, 1L, 1L, 1))));
+        this.mockMvc = MockMvcBuilders.standaloneSetup(tableRestController).build();
+        주문_테이블_응답_1 = new OrderTableResponse(1L, 1L, 5, false);
     }
 
     @Test
     void post() throws Exception {
         // given
-        given(orderService.create(any())).willReturn(주문_응답);
+        given(tableService.create(any())).willReturn(주문_테이블_응답_1);
 
         // when & then
         mockMvc.perform(MockMvcRequestBuilders.post(URI)
                         .contentType(APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(주문_요청)))
+                        .content(objectMapper.writeValueAsString(주문_테이블_응답_1)))
                 .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.orderStatus").value("COOKING"))
+                .andExpect(jsonPath("$.numberOfGuests").value(5))
                 .andDo(print());
     }
 
     @Test
     void get() throws Exception {
         // given
-        given(orderService.list()).willReturn(Collections.singletonList(주문_응답));
+        given(tableService.list()).willReturn(Collections.singletonList(주문_테이블_응답_1));
 
         // when & then
         mockMvc.perform(MockMvcRequestBuilders.get(URI)
                         .contentType(APPLICATION_JSON))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$[0].orderStatus").value("COOKING"))
+                .andExpect(jsonPath("$.length()").value(1))
                 .andDo(print());
     }
 
     @Test
-    void put() throws Exception {
+    void putForEmpty() throws Exception {
         // given
-        주문_응답 = new OrderResponse(1L, 1L, OrderStatus.COMPLETION.name(), null,
-                new ArrayList<>(Arrays.asList(new OrderLineItemResponse(1L, 1L, 1L, 1))));
-        given(orderService.changeOrderStatus(anyLong(), any())).willReturn(주문_응답);
+        given(tableService.changeEmpty(anyLong(), any())).willReturn(주문_테이블_응답_1);
 
         // when & then
-        mockMvc.perform(MockMvcRequestBuilders.put(URI + "/{orderId}/order-status", 1)
+        mockMvc.perform(MockMvcRequestBuilders.put(URI + "/{orderTableId}/empty", 1)
                         .contentType(APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsString(new OrderStatusRequest(OrderStatus.COMPLETION))))
+                        .content(objectMapper.writeValueAsString(주문_테이블_응답_1)))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.orderStatus").value("COMPLETION"))
+                .andDo(print());
+    }
+
+
+    @Test
+    void putForNumberOfGuests() throws Exception {
+        // given
+        given(tableService.changeNumberOfGuests(anyLong(), any())).willReturn(주문_테이블_응답_1);
+
+        // when & then
+        mockMvc.perform(MockMvcRequestBuilders.put(URI + "/{orderTableId}/number-of-guests", 1)
+                        .contentType(APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(주문_테이블_응답_1)))
+                .andExpect(status().isOk())
                 .andDo(print());
     }
 }
